Update poissondistribution.tex
When in the solutions we find P(N(0, h] = 0) = Poisson(lambda*h),  we are working with a Poisson distribution with parameter lambda*h and k=0. We end up with an exp{-lambda*h}. Clearly then, by Taylor expansion for abs(lambda*h) << 1, this converges to 1 - lambda*h + o(lambda*h)  (if I understand the o(h) notation correctly). If only abs(h) << 1, then for large lambda, we might still have that abs(lambda*h) is not << 1. In this case, we cannot apply the Taylor expansion in the way that it is applied in the solutions. Thus I added the line 'such that $|\lambda h| \ll 1$' and changed 'o(h)' to 'o(lambda*h)', as we are working with a function of lambda*h and not a function of just h. Furthermore, I changed the odd looking 1 in the solutions to the probability P letter since I think that's what it was supposed to be. I might just have missunderstood the solution to the exercise though! Please take a good look.