Update expdistribution.tex
Maybe a bit of a stupid adjustment, but shouldn't this be 'have seen'? Since jobs can 'depart', the expected time to see i jobs is not the same as the expected time to have seen i jobs.  And even if jobs can't depart yet, since the text hasn't spoken of it so far, this sentencing is definitely not worse than the sentencing that was used previously. It's at least greater than or equal to. That is, if I understand the question correctly.