Update constructiondiscretetime.tex
Do not approve this pull request.

When reading the queueing_book_mobile.pdf, the text which I put in between hyphens falls off the page. The text reads 'it must also be the number the start of period k.' I don't know if this is a general problem or just a problem for me, please take a look.